+++ tests/test_utils.py
@@ -4,11 +4,13 @@ from unittest.mock import call, patch
 
 import pytest
 from botocore.stub import ANY, Stubber
+from moto import mock_aws
+from moto.core import patch_client
 
 import opera_disp_tms.utils as ut
 
 
-@pytest.fixture(autouse=True)
+@pytest.fixture(autouse=False)
 def s3_stubber():
     with Stubber(ut.S3_CLIENT) as stubber:
         yield stubber
@@ -58,3 +60,116 @@ def test_upload_dir_to_s3(tmp_path):
                 call(tmp_path / 'subdir1/subdir3/bar.txt', 'myBucket', 'myPrefix/subdir1/subdir3/bar.txt'),
             ]
         )
+
+
+@mock_aws
+def test_list_files_in_s3(s3_bucket):
+    prefix = 'geotiffs'
+    geotiffs = [
+        f'{prefix}/my-file1.tif',
+        f'{prefix}/my-file2.tif',
+        f'{prefix}/my-file3.tif',
+    ]
+
+    for tif in geotiffs:
+        ut.S3_CLIENT.put_object(Bucket=s3_bucket, Key=tif)
+
+    files = ut.list_files_in_s3(s3_bucket, prefix)
+
+    assert len(files) == len(geotiffs)
+    assert files[0]['Key'].endswith('.tif')
+
+
+@mock_aws
+def test_download_file_from_s3(tmp_path, s3_bucket):
+    download_path = 'geotiffs/my-file.tif'
+    dest_dir = Path(tmp_path)
+
+    ut.S3_CLIENT.put_object(Bucket=s3_bucket, Key=download_path)
+
+    output_path = ut.download_file_from_s3(s3_bucket, download_path, dest_dir)
+
+    assert output_path == tmp_path / 'my-file.tif'
+
+
+@pytest.fixture
+def s3_bucket(scope='function'):
+    with mock_aws():
+        patch_client(ut.S3_CLIENT)
+
+        bucketName = 'myBucket'
+        location = {'LocationConstraint': 'us-west-2'}
+
+        ut.S3_CLIENT.create_bucket(Bucket=bucketName, CreateBucketConfiguration=location)
+
+        yield bucketName
+
+
+def test_get_frame_id():
+    info = {
+        'metadata': {
+            '': {
+                'frame_id': '11113',
+            },
+        },
+    }
+    with patch('osgeo.gdal.Info', return_value=info) as mock_info:
+        assert ut.get_frame_id('foo.tiff') == 11113
+        mock_info.assert_called_once_with('foo.tiff', format='json')
+
+    info = {
+        'metadata': {
+            '': {
+                'frame_id': '321',
+            },
+        },
+    }
+    with patch('osgeo.gdal.Info', return_value=info) as mock_info:
+        assert ut.get_frame_id('bar.tiff') == 321
+        mock_info.assert_called_once_with('bar.tiff', format='json')
+
+
+def test_get_west_most_point():
+    info = {
+        'wgs84Extent': {
+            'type': 'Polygon',
+            'coordinates': [
+                [
+                    [-122.201872, 41.1562811],
+                    [-122.201872, 39.2490859],
+                    [-118.7684233, 39.2490859],
+                    [-118.7684233, 41.1562811],
+                    [-122.201872, 41.1562811],
+                ]
+            ],
+        },
+    }
+    with patch('osgeo.gdal.Info', return_value=info) as mock_info:
+        assert ut.get_west_most_point('foo.tiff') == -122.201872
+        mock_info.assert_called_once_with('foo.tiff', format='json')
+
+    info = {
+        'wgs84Extent': {
+            'type': 'Polygon',
+            'coordinates': [
+                [
+                    [176.4143837, 53.0448494],
+                    [176.4418596, 50.8863044],
+                    [-179.4626094, 50.8340006],
+                    [-179.2889357, 52.9883554],
+                    [176.4143837, 53.0448494],
+                ]
+            ],
+        },
+    }
+    with patch('osgeo.gdal.Info', return_value=info) as mock_info:
+        assert ut.get_west_most_point('bar.tiff') == -179.4626094
+        mock_info.assert_called_once_with('bar.tiff', format='json')
+
+
+def test_get_common_direction():
+    assert ut.get_common_direction({1, 2, 3}) == 'ASCENDING'
+    assert ut.get_common_direction({1, 873, 46986}) == 'ASCENDING'
+    assert ut.get_common_direction({68, 69, 70}) == 'DESCENDING'
+    with pytest.raises(ValueError):
+        ut.get_common_direction({67, 68})